Move code into QuestionnaireList.vue
